v2/requirements.txt
pandas
psutil
requests
opencv-python
numpy
av
pyzbar
Flask
flask-restx
pyserial
werkzeug